refactor: move tests to __tests__ folder
