+++ src/components/__mocks__/seo.tsx
@@ -1,0 +1,5 @@
+import React, { ReactElement } from 'react';
+
+const SEO = (): ReactElement => <></>;
+
+export { SEO };
